feat: Prompt for SSH credentials to copy SSH keys
Users are now only required to provide a list of hosts in the Ansible
inventory file ('software_hosts'). If no SSH credentials are provided
then the user will be prompted to 1) create or select an SSH key pair
and 2) provide a username/password (via cli prompts) in order to copy
the SSH public key to all hosts. The Ansible inventory is updated with
the appropriate vars to continue using the selected SSH key. The
interactive configuration only works with a single global SSH user name
and password, but a manually configuring host specific variables in the
Ansible inventory should still work.

The inventory file ('software_hosts') is moved to the
'power-up/software' directory, and it is owned by the "login" user
instead of root.

SSH 'StrictHostKeyChecking' is forced to be disabled.

+++ scripts/python/software_hosts.py
@@ -20,7 +20,7 @@ from __future__ import nested_scopes, generators, division, absolute_import, \
 
 import click
 import os.path
-from os import listdir, mkdir, chown, chmod, getlogin
+from os import listdir, mkdir, chown, chmod, getlogin, getuid
 import filecmp
 import json
 import pwd
@@ -37,9 +37,9 @@ from inventory import generate_dynamic_inventory
 from lib.exception import UserException
 import lib.logger as logger
 from lib.genesis import get_python_path, CFG_FILE, \
-    get_dynamic_inventory_path, get_playbooks_path, get_ansible_path
+    get_dynamic_inventory_path, GEN_SOFTWARE_PATH, get_ansible_path
 from lib.utilities import bash_cmd, sub_proc_exec, heading1, get_selection, \
-    bold, get_yesno, sub_proc_display, remove_line, append_line
+    bold, get_yesno, sub_proc_display, remove_line, append_line, rlinput
 
 
 def _get_dynamic_inventory():
@@ -177,34 +177,26 @@ def _create_new_software_inventory(software_hosts_file_path):
 # For detailed information visit:
 #   http://docs.ansible.com/ansible/latest/user_guide/intro_inventory.html
 #
-# Global SSH logic credentials can be specified with [all:vars]
-#   e.g.:
-#   [all:vars]
-#   ansible_ssh_private_key_file=~/.ssh/powerup
-#   ansible_user=root
-#   ansible_ssh_common_args='-o StrictHostKeyChecking=no'
+# Only host definitions are required. SSH keys can be automatically
+# configured by pup or manually defined in this file.
 #
 # Group names are defined within brackets
 # Hosts are defined with an FQDN or IP address
 #   e.g.:
 #   [master]
 #   host1.domain.com  # master host1
-#   192.168.1.21      # master host2
+#   host2.domain.com  # master host2
 #
 #   [worker]
 #   host3.domain.com  # worker host1
-#   192.168.1.22      # worker host2
-#
-# Localhost can be specified as:
-#   localhost ansible_connection=local
-
-[all:vars]
+#   host4.domain.com  # worker host2
 
 [master]
   # define first master host on this line before the "#"
 
 [worker]
   # define first worker host on this line before the "#"
+
 """)
     hosts = None
     while hosts is None:
@@ -214,7 +206,24 @@ def _create_new_software_inventory(software_hosts_file_path):
                 new_hosts_file.write(hosts)
         elif not click.confirm('File not written! Try again?'):
             return False
+
+    _set_software_hosts_owner_mode(software_hosts_file_path)
+
     return True
+
+
+def _set_software_hosts_owner_mode(software_hosts_file_path):
+    """Set software_hosts file owner to "login" user
+
+    Args:
+        software_hosts_file_path (str): Path to software inventory file
+    """
+    user_name = getlogin()
+    if getuid() == 0 and user_name != 'root':
+        user_uid = pwd.getpwnam(user_name).pw_uid
+        user_gid = grp.getgrnam(user_name).gr_gid
+        os.chown(software_hosts_file_path, user_uid, user_gid)
+        os.chmod(software_hosts_file_path, 0o644)
 
 
 def _validate_inventory_count(software_hosts_file_path):
@@ -333,6 +342,7 @@ def configure_ssh_keys(software_hosts_file_path):
     Args:
         software_hosts_file_path (str): Path to software inventory file
     """
+    log = logger.getlogger()
     default_ssh_key_name = "powerup"
 
     ssh_key_options = get_existing_ssh_key_pairs()
@@ -352,33 +362,88 @@ def configure_ssh_keys(software_hosts_file_path):
     else:
         ssh_key = item
 
-    copy_ssh_key_pair_to_user(ssh_key)
+    copy_ssh_key_pair_to_user_dir(ssh_key)
 
-    hosts_list = _validate_inventory_count(software_hosts_file_path)
+    add_software_hosts_global_var(
+        software_hosts_file_path,
+        "ansible_ssh_common_args='-o StrictHostKeyChecking=no'")
 
-    cmd = (f'ansible-inventory --inventory {software_hosts_file_path} --list')
-    resp, err, rc = sub_proc_exec(cmd, shell=True)
-    hostvars = json.loads(resp)['_meta']['hostvars']
+    hostvars = get_ansible_hostvars(software_hosts_file_path)
+    global_user = None
+    global_pass = None
+    for host in _validate_inventory_count(software_hosts_file_path):
+        if global_user is None and 'ansible_user' not in hostvars[host]:
+            print(bold('\nOne or more hosts does not have \'ansible_user\' '
+                       'defined in inventory'))
+            global_user = rlinput('Please input a global client SSH login '
+                                  'username: ')
+            add_software_hosts_global_var(software_hosts_file_path,
+                                          f'ansible_user={global_user}')
+        if global_pass is None and 'ansible_ssh_pass' not in hostvars[host]:
+            print(bold('\nOne or more hosts does not have '
+                       '\'ansible_ssh_pass\' defined in inventory'))
+            global_pass = rlinput('Please input a global client SSH login '
+                                  'password: ')
+        if global_user is not None and global_pass is not None:
+            break
 
-    for host in hosts_list:
-        cmd = (f'ssh-copy-id {hostvars[host]["ansible_user"]}@{host} '
-               f'-i {ssh_key} ')
-        if "ansible_port" in hostvars[host]:
-            cmd += f'-p {hostvars[host]["ansible_port"]} '
-        if "ansible_ssh_common_args" in hostvars[host]:
-            cmd += f'{hostvars[host]["ansible_ssh_common_args"]} '
+    run = True
+    while run:
+        heading1("Copying SSH Public Keys to Hosts\n")
+        rc = copy_ssh_key_pair_to_hosts(ssh_key, software_hosts_file_path,
+                                        global_pass)
+        if not rc:
+            log.warning("One or more SSH key copy failed!")
+            choice, item = get_selection(['Retry', 'Continue', 'Exit'])
+            if choice == "1":
+                pass
+            elif choice == "2":
+                run = False
+            elif choice == "3":
+                log.debug('User chooses to exit.')
+                sys.exit('Exiting')
+        else:
+            log.info("SSH key successfully copied to all hosts")
+            run = False
 
-        rc = sub_proc_display(cmd)
+    add_software_hosts_global_var(software_hosts_file_path,
+                                  f'ansible_ssh_private_key_file={ssh_key}')
 
+
+def add_software_hosts_global_var(software_hosts_file_path, entry):
+    """Copy an SSH public key into software hosts authorized_keys files
+
+    Add entry to software_hosts '[all:vars]' section. Any existing
+    entries with the same key name (string before '=') will be
+    overwritten.
+
+    Args:
+        software_hosts_file_path (str): Path to software inventory file
+        entry (str) : Entry to write in software_hosts '[all:vars]'
+    """
     remove_line(software_hosts_file_path, '^ansible_ssh_private_key_file=.*')
+
     append_line(software_hosts_file_path, '[all:vars]')
+
     with open(software_hosts_file_path, 'r') as software_hosts_read:
         software_hosts = software_hosts_read.readlines()
+
+    in_all_vars = False
+    prev_line = "BOF"
     with open(software_hosts_file_path, 'w') as software_hosts_write:
         for line in software_hosts:
             if line.startswith("[all:vars]"):
-                line = line + f'ansible_ssh_private_key_file={ssh_key}\n'
+                if prev_line != "\n":
+                    line = "\n" + line
+                line = line + f'{entry}\n'
+                in_all_vars = True
+            elif in_all_vars and line.startswith('['):
+                in_all_vars = False
+            elif in_all_vars and line.startswith(entry.split('=')[0]):
+                continue
             software_hosts_write.write(line)
+            prev_line = line
+    _set_software_hosts_owner_mode(software_hosts_file_path)
 
 
 def get_existing_ssh_key_pairs():
@@ -433,7 +498,7 @@ def create_ssh_key_pair(name):
     if os.path.isfile(private_key_path):
         log.info(f'SSH key \'{private_key_path}\' already exists, continuing')
     else:
-        log.info(f'Creating SSH key \'{private_key_path}\'')
+        print(bold(f'Creating SSH key \'{private_key_path}\''))
         cmd = ('ssh-keygen -t rsa -b 4096 '
                '-C "Generated by Power-Up Software Installer" '
                f'-f {private_key_path} -N ""')
@@ -445,7 +510,7 @@ def create_ssh_key_pair(name):
     return private_key_path
 
 
-def copy_ssh_key_pair_to_user(private_key_path):
+def copy_ssh_key_pair_to_user_dir(private_key_path):
     """Copy an SSH private/public key pair into the user's ~/.ssh dir
 
     This function is useful when a key pair is created as root user
@@ -476,23 +541,85 @@ def copy_ssh_key_pair_to_user(private_key_path):
             os.chown(user_ssh_dir, user_uid, user_gid)
 
         # Never overwrite an existing private key file!
+        already_copied = False
         while os.path.isfile(user_private_key_path):
             # If key pair already exists no need to do anything
             if (filecmp.cmp(private_key_path, user_private_key_path) and
                     filecmp.cmp(public_key_path, user_public_key_path)):
-                return
+                already_copied = True
+                break
             else:
                 user_private_key_path += "_powerup"
                 user_public_key_path = user_private_key_path + '.pub'
 
-        copyfile(private_key_path, user_private_key_path)
-        copyfile(public_key_path, user_public_key_path)
+        if already_copied:
+            print(f'\'{private_key_path}\' already copied to '
+                  f'\'{user_private_key_path}\'')
+        else:
+            print(bold(f'Copying \'{private_key_path}\' to '
+                       f'\'{user_private_key_path}\' for unprivileged use'))
+            copyfile(private_key_path, user_private_key_path)
+            copyfile(public_key_path, user_public_key_path)
 
-        os.chown(user_private_key_path, user_uid, user_gid)
-        os.chmod(user_private_key_path, 0o600)
+            os.chown(user_private_key_path, user_uid, user_gid)
+            os.chmod(user_private_key_path, 0o600)
 
-        os.chown(user_public_key_path, user_uid, user_gid)
-        os.chmod(user_public_key_path, 0o644)
+            os.chown(user_public_key_path, user_uid, user_gid)
+            os.chmod(user_public_key_path, 0o644)
+
+
+def copy_ssh_key_pair_to_hosts(private_key_path, software_hosts_file_path,
+                               global_pass=None):
+    """Copy an SSH public key into software hosts authorized_keys files
+
+    TODO: detailed description
+
+    Args:
+        private_key_path (str) : Filename of private key file
+        software_hosts_file_path (str): Path to software inventory file
+        global_pass (str, optional): Global client default SSH password
+
+    Returns:
+        bool: True iff rc of all commands are "0"
+    """
+    log = logger.getlogger()
+    hosts_list = _validate_inventory_count(software_hosts_file_path)
+    all_zero_returns = True
+
+    hostvars = get_ansible_hostvars(software_hosts_file_path)
+
+    for host in hosts_list:
+        print(bold(f'Copy SSH Public Key to {host}'))
+        cmd = f'ssh-copy-id -i {private_key_path} '
+        if "ansible_port" in hostvars[host]:
+            cmd += f'-p {hostvars[host]["ansible_port"]} '
+        if "ansible_ssh_common_args" in hostvars[host]:
+            cmd += f'{hostvars[host]["ansible_ssh_common_args"]} '
+        cmd += f'{hostvars[host]["ansible_user"]}@{host}'
+
+        if 'ansible_ssh_pass' not in hostvars[host]:
+            cmd = f'SSHPASS=\'{global_pass}\' sshpass -e ' + cmd
+
+        rc = sub_proc_display(cmd, shell=True)
+        if rc != 0:
+            all_zero_returns = False
+
+    return all_zero_returns
+
+
+def get_ansible_hostvars(software_hosts_file_path):
+    """Get Ansible generated 'hostvars' dictionary
+
+    Args:
+        software_hosts_file_path (str): Path to software inventory file
+
+    Returns:
+        dict: Ansible 'hostvars' dictionary
+    """
+    cmd = (f'ansible-inventory --inventory {software_hosts_file_path} --list')
+    resp, err, rc = sub_proc_exec(cmd, shell=True)
+    hostvars = json.loads(resp)['_meta']['hostvars']
+    return hostvars
 
 
 def get_user_and_home():
@@ -559,7 +686,7 @@ def get_ansible_inventory():
     inventory_choice = None
     dynamic_inventory_path = get_dynamic_inventory_path()
     software_hosts_file_path = (
-        os.path.join(get_playbooks_path(), 'software_hosts'))
+        os.path.join(GEN_SOFTWARE_PATH, 'software_hosts'))
 
     heading1("Software hosts inventory setup\n")
 
@@ -590,7 +717,8 @@ def get_ansible_inventory():
             # If no software inventory file exists create one using template
             else:
                 print("Software inventory file not found.")
-                if click.confirm('Do you want to create a new inventory from a template?'):
+                if click.confirm('Do you want to create a new inventory from '
+                                 'a template?'):
                     _create_new_software_inventory(software_hosts_file_path)
                 elif click.confirm('Do you want to exit the program?'):
                     sys.exit(1)
